fix: Codex 리뷰 P1/P2 반영 — 순방향 마이그레이션 분리 및 unsigned 모터 카운터 오버플로 수정
[P1] V1__init_schema.sql을 원상 복구하고, tb_data_rcv_anal 8개 컬럼 타입 정합화를
V2__fix_tb_data_rcv_anal_numeric_column_types.sql로 분리했다. 이미 V1이 적용된 환경에서
V1 자체를 고치면 Flyway 체크섬 검증이 깨져 기동이 막히기 때문이다. V2는
"BASELINE DB 자동 ALTER 금지 원칙"(docs/flyway-migration-recovery.md)에 따라
flyway_schema_history에 V1=BASELINE 행이 있는 환경(=이미 dev 실측 타입을 가진 공유
레거시 DB)에서는 스스로 건너뛰고, 신규 DB(V1의 CREATE TABLE이 실제로 실행된 환경)에서만
ALTER를 적용한다.

[P2] GateStatusAnalyzer의 motorCount/masterInTotal이 unsigned 32비트 카운터(최대
4,294,967,295)인데도 u32(...).toInt()로 좁혀 값이 음수로 래핑될 수 있었다. Long으로
유지하도록 LaneStatusAnalysis.motorCount/masterInTotal, DataReceiveAnalysis.descMotorCount,
GateStatusAnalysisPersister/OprStatusPersister의 관련 코드를 함께 수정했다
(masterInTotal은 P2와 동일한 근본 원인이라 함께 고쳤다).

전체 컴파일 및 securance-protocol/domain/server/web 테스트 통과 확인
(GateTcpClientTest의 1건 실패는 재실행 시 통과하는 무관한 소켓 타이밍 flake).

+++ securance-domain/src/main/kotlin/kr/co/securance/secuhub/domain/entity/DataReceiveAnalysis.kt
@@ -224,8 +224,10 @@ class DataReceiveAnalysis(
     @Column(name = "desc_operation08", nullable = false, length = 20)
     var descOperation08: String = "",
 
+    // unsigned 32비트 카운터(최대 4,294,967,295)라 Int로는 범위를 다 담지 못한다(Codex 리뷰 P2,
+    // 2026-09-04) — Long으로 매핑한다. DB 컬럼 자체는 dev DB 실측대로 int(10) unsigned 그대로 둔다.
     @Column(name = "desc_motor_count", nullable = false)
-    var descMotorCount: Int = 0,
+    var descMotorCount: Long = 0,
 
     @Column(name = "desc_master_in_total", nullable = false)
     var descMasterInTotal: Long = 0,
